+++ resources/views/user/view-ticket.blade.php
@@ -57,20 +57,17 @@
     <div class="card-body" id="formshoworhide">
         <form id="AddCommentForm" class="needs-validation" novalidate>
             <div class="row">
-                <input type="hidden" id="ticket_user_Id">
-                <input type="hidden" id="ticket_Id">
+                <input type="hidden" id="comment_ticket_Id">
             <div class="col-md-12 mt-2">
                 <label for="validationCustom01">Comment<span class="text-danger">*</span></label>
                 <textarea required class="form-control" id="comment_Text" rows="5" placeholder="Enter your response here."></textarea>
             </div>
-            <div class="row d-flex justify-content-end">
-                <div class="col-md-2 mt-2">
-                    <button type="submit" class="btn btn-primary col-xl-12" id="comment_Submit">Submit</button>
+            <div class="row">
+                <div class="col-12 mt-2 d-flex justify-content-end">
+                    <a class="btn btn-info m-1" data-bs-toggle="offcanvas" href="#trailOffCanvas" role="button" aria-controls="offcanvasEnd">View trail</a>
+                    <button type="submit" class="btn btn-primary m-1" id="comment_Submit">Submit</button>
                 </div>
         </form>
-                <div class="col-md-2 mt-2">
-                    <a class="btn btn-info col-xl-12" data-bs-toggle="offcanvas" href="#trailOffCanvas" role="button" aria-controls="offcanvasEnd">View trail</a>
-                </div>
             </div>
     </div>      
 </div>
@@ -94,6 +91,7 @@
   @endslot
 
   @slot('customJS')
+  <script src="{{ asset('libs/bootstrap/validation.js')}}" defer></script>
   <script src="{{ asset('js/ajax/user/view-ticket.ajax.js')}}" defer></script>
   @endslot
 </x-user-layout>